Dump function types as they are, without adhoc tricks.
SDCC dumps all types as they are, and only function types suddenly
dumped with adhoc "function" sticking into. This will break parsing
of such type (essentially, it's not a type, but a declaration).
Hopefully, folks who look into compiler representation know that
there's function type in C, it's written like "int()" (function),
and the fact that it's promoted to a pointer to a function "int (*)()"
is separate issue.